Intent: Consolidate IO coherence across structured serialization Operators involved: [Coupling, Coherence] Affected invariants: [#4, #7] Key changes: - Move JSON serialization helpers into tnfr.io and document consolidated API with stubs - Provide tnfr.utils.io compatibility shim and update internal imports/tests to new entry point - Refresh typing exports and property/unit tests to consume tnfr.io.json_dumps directly Expected risks/dissonances: Minimal risk of lingering deprecated imports; mitigated with compatibility shim and updated tests Metrics: Expect consistent JSON round-trips and unchanged structured file parsing telemetry Equivalence map: tnfr.utils.io.json_dumps -> tnfr.io.json_dumps
