+++ src/tnfr/cli/execution.py
@@ -38,7 +38,8 @@ from ..ontosim import prepare_network
 from ..sense import register_sigma_callback
 from ..trace import register_trace
 from ..types import ProgramTokens
-from ..utils import get_logger, json_dumps
+from ..io import json_dumps
+from ..utils import get_logger
 from .arguments import _args_to_dict
 
 logger = get_logger(__name__)
