+++ src/tnfr/gamma.py
@@ -12,13 +12,13 @@ from typing import Any, Callable, NamedTuple
 
 from .alias import get_theta_attr
 from .constants import DEFAULTS
+from .io import json_dumps
 from .metrics.trig_cache import get_trig_cache
 from .types import GammaSpec, NodeId, TNFRGraph
 from .utils import (
     edge_version_cache,
     get_graph_mapping,
     get_logger,
-    json_dumps,
     node_set_checksum,
 )
