+++ src/tnfr/metrics/export.py
@@ -10,9 +10,8 @@ from typing import Mapping, TextIO
 
 from ..config.constants import GLYPHS_CANONICAL
 from ..glyph_history import ensure_history
-from ..io import safe_write
+from ..io import json_dumps, safe_write
 from ..types import Graph
-from ..utils import json_dumps
 from .core import glyphogram_series
 from .glyph_timing import SigmaTrace
