+++ tests/unit/structural/test_json_utils.py
@@ -2,8 +2,8 @@ import logging
 
 import pytest
 
+import tnfr.io as json_utils
 import tnfr.utils.init as import_utils
-import tnfr.utils.io as json_utils
 
 from ...utils import clear_orjson_cache
