+++ tests/utils.py
@@ -1,4 +1,4 @@
-import tnfr.utils.io as json_utils
+import tnfr.io as json_utils
 
 
 def clear_orjson_cache() -> None:
